Use last_downstream_green instead of last_green
> last_green refers to the Bazel binary that was built at the most recent commit that passed Bazel CI. Ideally this binary should be very close to Bazel-at-head.
> last_downstream_green points to the most recent Bazel binary that builds and tests all downstream projects successfully.

We should not blindly use Bazel@HEAD for Tulsi, instead we should use a known commit that works for all downstream projects (including Tulsi). This way, a bug or an incompatible change like
bazelbuild/bazel#10829 can be detected when we see Tulsi is green in the main pipeline but red in the downstream pipeline.

Close https://github.com/bazelbuild/tulsi/pull/129

PiperOrigin-RevId: 296914365