`expand_location` of copts
This change simply checks for `$(` and then calls `ctx.expand_copt`.
This also makes tulsi work with other community rule sets that use this
`$(execpath` or other bazel make variables in `copts`.

We just landed a series of changes to build the top 50 swift cocoapods
with PodToBUILD. Like other usages of building mixed module swift libs,
it uses `$(execpath` to pass an internal module map via
`fmodule-map-file=`. This is required so it doesn't propagate upwards.
https://github.com/pinterest/PodToBUILD/pull/153

Change-Id: I457b2ac15025b792a1a4a90511a7ec98a83bfe26