Tulsi-generated Xcode projects will now use a bundled lldbinit that loads the global lldbinit and then applies project specific settings only to debug sessions in the given Xcode project. This feature is only available in Xcode >= 11.4 so this change retains the previous behavior of using the default lldbinit, i.e. the global lldbinit, and modifying it to load a global Tulsi-specific lldbinit. This previous approach leaks Tulsi-specific settings to all instances of lldb.
An option, "Disable project-level lldbinit", is also added to manually revert to the previous behavior if issues arise.

This change also adds an additional setting to the lldbinit (global or project-specific) to set the LLDB working directory to the bazel workspace root. Doing so will allow lldb to find files whose paths are stored relative to the bazel root.

PiperOrigin-RevId: 345752843