Add an alternate codepath to BootstrapLLDBInit that undoes the lldbinit linking.
This codepath should be taken when using a project specific lldbinit. This is needed when debugging UI tests because the test host does not seem to load the project level lldbinit, instead it loads the global lldbinit which can include a reference to ~/.lldbinit-tulsiproj left over from previous Tulsi projects. The test host appears to inherit the lldb settings from the test runner so our project lldbinit will still apply to the test host as long as an ~/.lldbinit-tulsiproj does not override them.

PiperOrigin-RevId: 362540211