Fix an issue where the lldbinit parsing logic throws an out of range error when the lldbinit contained content after the tulsi block.
The source_idx var which tracks which line in the tulsi block we are looking for was being incremented past the length of the tulsi block resulting in an out of range error. Once source_idx exceeds the length of the tulsi block, it no longer makes sense to check whether the line in the lldbinit file matches a line in the tulsi block so this change adds a gaurd for that condition.

PiperOrigin-RevId: 363674993
(cherry picked from commit 16d7b338e2dffb2e53a766e08bcfd0ca0a8135b8)