Migrate away from "deps" to "storyboards" and "resources".
"deps" implies code dependencies, which the existing watchos_application rule does not support. However, single target watchOS applications which will be supported by a watchos_single_target_application rule and eventually the watchos_application rule itself do own code dependencies. That changes the meaning of "deps" as it is used today.

The "deps" attr will be removed from the watchos_application in a future change on rules_apple, which will be tracked in [].

PiperOrigin-RevId: 474890643